Moving last diffusion tests from v3 to v4
And also moving BiLinear interpolation to interpolation.py
