+++ tests/v4/test_diffusion.py
@@ -6,36 +6,13 @@ from scipy import stats
 
 from parcels._datasets.structured.generic import simple_UV_dataset
 from parcels.application_kernels import AdvectionDiffusionEM, AdvectionDiffusionM1, DiffusionUniformKh
+from parcels.application_kernels.interpolation import XBiLinear
 from parcels.field import Field, VectorField
 from parcels.fieldset import FieldSet
+from parcels.particle import Particle, Variable
 from parcels.particleset import ParticleSet
-from parcels.xgrid import _XGRID_AXES, XGrid
-
-
-def BiLinear(  # TODO move to interpolation file
-    field: Field,
-    ti: int,
-    position: dict[_XGRID_AXES, tuple[int, float | np.ndarray]],
-    tau: np.float32 | np.float64,
-    t: np.float32 | np.float64,
-    z: np.float32 | np.float64,
-    y: np.float32 | np.float64,
-    x: np.float32 | np.float64,
-):
-    """Bilinear interpolation on a regular grid."""
-    xi, xsi = position["X"]
-    yi, eta = position["Y"]
-    zi, zeta = position["Z"]
-
-    data = field.data.data[:, zi, yi : yi + 2, xi : xi + 2]
-    data = (1 - tau) * data[ti, :, :] + tau * data[ti + 1, :, :]
-
-    return (
-        (1 - xsi) * (1 - eta) * data[0, 0]
-        + xsi * (1 - eta) * data[0, 1]
-        + xsi * eta * data[1, 1]
-        + (1 - xsi) * eta * data[1, 0]
-    )
+from parcels.xgrid import XGrid
+from tests.utils import create_fieldset_zeros_conversion
 
 
 @pytest.mark.parametrize("mesh_type", ["spherical", "flat"])
@@ -48,12 +25,12 @@ def test_fieldKh_Brownian(mesh_type):
     ds["lon"].data = np.array([-1e6, 1e6])
     ds["lat"].data = np.array([-1e6, 1e6])
     grid = XGrid.from_dataset(ds)
-    U = Field("U", ds["U"], grid, mesh_type=mesh_type, interp_method=BiLinear)
-    V = Field("V", ds["V"], grid, mesh_type=mesh_type, interp_method=BiLinear)
+    U = Field("U", ds["U"], grid, mesh_type=mesh_type, interp_method=XBiLinear)
+    V = Field("V", ds["V"], grid, mesh_type=mesh_type, interp_method=XBiLinear)
     ds["Kh_zonal"] = (["time", "depth", "YG", "XG"], np.full((2, 1, 2, 2), kh_zonal))
     ds["Kh_meridional"] = (["time", "depth", "YG", "XG"], np.full((2, 1, 2, 2), kh_meridional))
-    Kh_zonal = Field("Kh_zonal", ds["Kh_zonal"], grid=grid, mesh_type=mesh_type, interp_method=BiLinear)
-    Kh_meridional = Field("Kh_meridional", ds["Kh_meridional"], grid=grid, mesh_type=mesh_type, interp_method=BiLinear)
+    Kh_zonal = Field("Kh_zonal", ds["Kh_zonal"], grid=grid, mesh_type=mesh_type, interp_method=XBiLinear)
+    Kh_meridional = Field("Kh_meridional", ds["Kh_meridional"], grid=grid, mesh_type=mesh_type, interp_method=XBiLinear)
     UV = VectorField("UV", U, V)
     fieldset = FieldSet([U, V, UV, Kh_zonal, Kh_meridional])
 
@@ -85,8 +62,8 @@ def test_fieldKh_SpatiallyVaryingDiffusion(mesh_type, kernel):
     ds["lon"].data = np.linspace(-1e6, 1e6, xdim)
     ds["lat"].data = np.linspace(-1e6, 1e6, ydim)
     grid = XGrid.from_dataset(ds)
-    U = Field("U", ds["U"], grid, mesh_type=mesh_type, interp_method=BiLinear)
-    V = Field("V", ds["V"], grid, mesh_type=mesh_type, interp_method=BiLinear)
+    U = Field("U", ds["U"], grid, mesh_type=mesh_type, interp_method=XBiLinear)
+    V = Field("V", ds["V"], grid, mesh_type=mesh_type, interp_method=XBiLinear)
 
     Kh = np.zeros((ydim, xdim), dtype=np.float32)
     for x in range(xdim):
@@ -94,8 +71,8 @@ def test_fieldKh_SpatiallyVaryingDiffusion(mesh_type, kernel):
 
     ds["Kh_zonal"] = (["time", "depth", "YG", "XG"], np.full((2, 1, ydim, xdim), Kh))
     ds["Kh_meridional"] = (["time", "depth", "YG", "XG"], np.full((2, 1, ydim, xdim), Kh))
-    Kh_zonal = Field("Kh_zonal", ds["Kh_zonal"], grid=grid, mesh_type=mesh_type, interp_method=BiLinear)
-    Kh_meridional = Field("Kh_meridional", ds["Kh_meridional"], grid=grid, mesh_type=mesh_type, interp_method=BiLinear)
+    Kh_zonal = Field("Kh_zonal", ds["Kh_zonal"], grid=grid, mesh_type=mesh_type, interp_method=XBiLinear)
+    Kh_meridional = Field("Kh_meridional", ds["Kh_meridional"], grid=grid, mesh_type=mesh_type, interp_method=XBiLinear)
     UV = VectorField("UV", U, V)
     fieldset = FieldSet([U, V, UV, Kh_zonal, Kh_meridional])
     fieldset.add_constant("dres", ds["lon"][1] - ds["lon"][0])
@@ -110,3 +87,58 @@ def test_fieldKh_SpatiallyVaryingDiffusion(mesh_type, kernel):
     assert np.allclose(np.mean(pset.lon), 0, atol=tol)
     assert np.allclose(np.mean(pset.lat), 0, atol=tol)
     assert stats.skew(pset.lon) > stats.skew(pset.lat)
+
+
+@pytest.mark.parametrize("lambd", [1, 5])
+def test_randomexponential(lambd):
+    fieldset = create_fieldset_zeros_conversion()
+    npart = 1000
+
+    # Rate parameter for random.expovariate
+    fieldset.lambd = lambd
+
+    # Set random seed
+    random.seed(1234)
+
+    pset = ParticleSet(fieldset=fieldset, lon=np.zeros(npart), lat=np.zeros(npart), depth=np.zeros(npart))
+
+    def vertical_randomexponential(particle, fieldset, time):  # pragma: no cover
+        # Kernel for random exponential variable in depth direction
+        particle.depth = random.expovariate(fieldset.lambd)
+
+    pset.execute(vertical_randomexponential, runtime=np.timedelta64(1, "s"), dt=np.timedelta64(1, "s"))
+
+    expected_mean = 1.0 / fieldset.lambd
+    assert np.allclose(np.mean(pset.depth), expected_mean, rtol=0.1)
+
+
+@pytest.mark.parametrize("mu", [0.8 * np.pi, np.pi])
+@pytest.mark.parametrize("kappa", [2, 4])
+def test_randomvonmises(mu, kappa):
+    npart = 10000
+    fieldset = create_fieldset_zeros_conversion()
+
+    # Parameters for random.vonmisesvariate
+    fieldset.mu = mu
+    fieldset.kappa = kappa
+
+    # Set random seed
+    random.seed(1234)
+
+    AngleParticle = Particle.add_variable(Variable("angle"))
+    pset = ParticleSet(
+        fieldset=fieldset, pclass=AngleParticle, lon=np.zeros(npart), lat=np.zeros(npart), depth=np.zeros(npart)
+    )
+
+    def vonmises(particle, fieldset, time):  # pragma: no cover
+        particle.angle = random.vonmisesvariate(fieldset.mu, fieldset.kappa)
+
+    pset.execute(vonmises, runtime=np.timedelta64(1, "s"), dt=np.timedelta64(1, "s"))
+
+    angles = np.array([p.angle for p in pset])
+
+    assert np.allclose(np.mean(angles), mu, atol=0.1)
+    vonmises_mean = stats.vonmises.mean(kappa=kappa, loc=mu)
+    assert np.allclose(np.mean(angles), vonmises_mean, atol=0.1)
+    vonmises_var = stats.vonmises.var(kappa=kappa, loc=mu)
+    assert np.allclose(np.var(angles), vonmises_var, atol=0.1)
